Move some more things, update documentation
