+++ src/main/background.js
@@ -1,3 +1,3 @@
-import './omnibar'
-import './menus-and-shortcuts'
-import './trigger-capture'
+import './background/omnibar'
+import './background/menus-and-shortcuts'
+import './background/trigger-capture'
