+++ src/main/background/omnibar.js
