+++ src/overview/overview-ui/reducer.js
