+++ src/overview/overview-ui/selectors.js
@@ -1,0 +1,4 @@
+// Gets this module's namespace within the full redux state (could be more elegant..)
+export function ourState(rootState) {
+    return rootState.overview
+}
