--- src/overview/selectors.js
@@ -1,5 +1,0 @@
-
-// Gets this module's namespace within the full redux state
-export function ourState(rootState) {
-    return rootState.overview
-}
