+++ src/page-analysis/content_script/index.js
@@ -1,8 +1,7 @@
-import { makeRemotelyCallable } from 'src/util/webextensionRPC'
-
-import extractPageContent from './extract-page-content'
 import freezeDry from 'freeze-dry'
 
+import { makeRemotelyCallable } from 'src/util/webextensionRPC'
+import { extractPageContent } from './extract-page-content'
 
 makeRemotelyCallable({
     extractPageContent,
